+++ base/src/main/java/io/github/sds100/keymapper/base/promode/ProModeSetupState.kt
@@ -9,4 +9,5 @@ data class ProModeSetupState(
     val stepContent: StepContent,
     val isSetupAssistantChecked: Boolean,
     val isSetupAssistantButtonEnabled: Boolean,
+    val isStarting: Boolean,
 )